update dashboard UI, seperate create meeting from create team, update team setting
